+++ src/controllers/ProductPurchaseManager.js
@@ -2,7 +2,7 @@ import { $, default as DOM } from '../views/DOMUtils.js';
 import { default as V } from '../utils/validators.js';
 import { default as DB } from '../model/database.js';
 import { default as UT } from '../utils/utils.js';
-import calculateReturnCoins from '../utils/calculateReturnCoins.js';
+import calculateReturnCoins from './utils/calculateReturnCoins.js';
 
 export default class ProductPurchaseManager {
   constructor() {
@@ -16,7 +16,7 @@ export default class ProductPurchaseManager {
   }
 
   manage() {
-    this.manageChargeToPurhcaseProduct;
+    this.manageChargeToPurhcaseProduct();
     this.manageAllPurchaseButton();
     this.manageCoinReturn();
   }
