+++ src/controllers/VendingMachineChargeManager.js
@@ -1,6 +1,6 @@
 import { $, default as DOM } from '../views/DOMUtils.js';
 import { default as V } from '../utils/validators.js';
-import convertChargeIntoCoin from '../utils/convertChargeIntoCoin.js';
+import convertChargeIntoCoin from './utils/convertChargeIntoCoin.js';
 
 export default class VendingMachineChargeManager {
   constructor() {
